+++ src/NovaPoshtaServiceProvider.php
@@ -17,9 +17,10 @@ class NovaPoshtaServiceProvider extends ServiceProvider
             __DIR__.'/../config/novaposhta.php' => config_path('novaposhta.php'),
         ], 'config');
 
-        $this->publishes([
-            __DIR__.'/../resources/lang' => $this->app['path.lang'].'/vendor/novaposhta',
-        ]);
+//        Сложно с обновлением, выключил
+//        $this->publishes([
+//            __DIR__.'/../resources/lang' => $this->app['path.lang'].'/vendor/novaposhta',
+//        ]);
 
         $this->loadTranslationsFrom(__DIR__.'/../resources/lang/', 'novaposhta');
     }